VERSION{v6.49}
It isn't often that someone annoys me so much that i actually end up naming a unit after them...
Meet the Kmar Shielding Unit!

It is a linkable shield that has 500 power and regenerates 20 power per second at an energy drain of 2. The Kmar replaces the unfortunately crippled "Thorn".

The Kmar is linkable, so you can build multipls close to one another in order to distribute their power. Their radius is pretty small though, and their hp is only 250.

In other news, I have regenerated all of the unit icons. Most of them should be much easier to distinguish.

git-svn-id: https://evolutionrts.googlecode.com/svn/trunk@625 a8037ef8-a490-e903-bf31-6b53b2d07c0b