VERSION{v6.54}
Rebalance air to make it less of a instawin item
Air unit costs halved
Air unit max damages halved
Air Factory cost halved (so it is exactly the same as the other factories now)
Bomber damage to light/armored raised to 50%, max damage halved (so in other words, does slightly more damage vs mobile units than it did, but less damage to buildings).
Gunship damage vs buildings raised to 25% so that it still does exactly the same amount of damage to buildings that it did (32.5), but it's damage vs light/armored units is lowered.
Fighter retains it's AA damage

git-svn-id: https://evolutionrts.googlecode.com/svn/trunk@636 a8037ef8-a490-e903-bf31-6b53b2d07c0b