Disabled message spam.
issue with the zone capture going by ally id. we need a way to get the ally color(which isn't a thing) so we are going to grab the 1st team, first player color and assign it to a table for the ally teams. That way we can reference this new table for the "ally color"

other areas of concern are a few k/v names in the code need cleaning up and proper names.

Still only show current player with lowest number in list instead of all with the TEAM and color above them all, work resumes tomorrow sometime!

tired for now but the code seems stable